Refactor card and table to common module
