+++ client/my-sites/plugins/plugin-management-v2/common/card/style.scss
@@ -1,38 +1,38 @@
-.plugin-card__columns {
+.card__columns {
 	display: flex;
   	flex:1;
 }
-.plugin-card__toggle-container {
+.card__toggle-container {
 	margin-block-start: 16px;
 	& > div {
 		margin: 16px 0;
 	}
 }
-.plugin-card__left-content {
+.card__left-content {
 	min-width: 32px;
 	max-width: 32px;
   	order: 1;
 }
-.plugin-card__main-content {
+.card__main-content {
 	flex: 1;
 	order: 2;
 	position: relative;
 	padding: 0 0 0 16px;
 	width: 70%;
 }
-.plugin-card__right-content {
+.card__right-content {
 	width: 20px;
 	order: 3;
 }
-.plugin-card__plugin-icon {
+.card__plugin-icon {
 	width: 32px;
 	height: 32px;
 }
-.plugin-card__all-actions {
+.card__all-actions {
 	color: var( --studio-gray-40 );
 	margin: 0 0.1em;
 }
-.plugin-card__overlay {
+.card__overlay {
 	display: block;
 	position: absolute;
 	height: 20px;
@@ -45,11 +45,11 @@
 	inset-block-start: 2px;
 	inset-inline-end: 0;
 }
-.plugin-card__site-data {
+.card__site-data {
 	color: var( --studio-gray-60 );
 	font-size: 0.75rem;
 	margin: 8px 0;
 }
-.plugin-card__update-plugin {
+.card__update-plugin {
 	margin-block-start: 16px;
 }
