+++ client/my-sites/plugins/plugin-management-v2/common/table/style.scss
@@ -2,7 +2,7 @@
 @import '@wordpress/base-styles/_breakpoints.scss';
 @import '@wordpress/base-styles/_mixins.scss';
 
-.plugins-table__table {
+.table__table {
 	display: none;
 	@include break-xlarge() {
 		display: table;
@@ -52,13 +52,13 @@
 		vertical-align: middle;
 	}
 }
-td.plugins-table__actions {
+td.table__actions {
 	padding: 0;
 	width: 50px;
 	text-align: right;
 	padding-right: 16px;
 }
-.plugins-table__table-row {
+.table__table-row {
 	&:hover {
 		background: var( --studio-gray-0 );
 		.plugin-row-formatter__overlay {
@@ -66,14 +66,14 @@ td.plugins-table__actions {
 		}
 	}
 }
-.plugins-table__all-actions {
+.table__all-actions {
 	vertical-align: middle;
 	display: inline-flex;
 	color: var( --studio-gray-40 );
 	margin: 0 0.1em;
 	height: 100%;
 }
-.plugins-table__small-column {
+.table__small-column {
 	max-width: 100px;
 	@include break-wide() {
 		max-width: 120px;
@@ -85,7 +85,7 @@ td.plugins-table__actions {
 .components-form-toggle.is-checked .components-form-toggle__track {
 	background-color: var( --studio-jetpack-green-50 );
 }
-.plugins-table__plugin-icon {
+.table__plugin-icon {
 	width: 24px;
 	height: 24px;
 	vertical-align: middle;
