+++ client/my-sites/plugins/plugin-management-v2/plugins-list/index.tsx
@@ -1,8 +1,9 @@
 import { Card } from '@automattic/components';
 import TextPlaceholder from 'calypso/jetpack-cloud/sections/partner-portal/text-placeholder';
 import PluginCard from '../plugin-card';
+import PluginRowFormatter from '../plugin-row-formatter';
 import PluginsTable from '../plugins-table';
-import type { PluginColumns, PluginRowFormatterArgs } from '../types';
+import type { Columns, PluginRowFormatterArgs, Plugin } from '../types';
 import type { SiteData } from 'calypso/state/ui/selectors/site-data';
 import type { ReactElement, ReactNode } from 'react';
 
@@ -10,39 +11,55 @@ import './style.scss';
 
 interface Props {
 	selectedSite: SiteData;
-	items: Array< any >;
+	items: Array< Plugin >;
 	isLoading: boolean;
-	columns: PluginColumns;
+	columns: Columns;
 	title?: ReactNode;
-	rowFormatter: ( args: PluginRowFormatterArgs ) => ReactNode;
 	hasMoreActions?: boolean;
 	primaryKey: string;
 }
 
 export default function PluginsList( {
+	items,
+	title,
+	isLoading,
 	hasMoreActions = true,
 	primaryKey,
+	selectedSite,
 	...rest
 }: Props ): ReactElement {
+	const rowFormatter = ( props: PluginRowFormatterArgs ) => {
+		return <PluginRowFormatter { ...props } selectedSite={ selectedSite } />;
+	};
+
 	return (
 		<>
-			<PluginsTable { ...rest } hasMoreActions={ hasMoreActions } primaryKey={ primaryKey } />
+			<PluginsTable
+				{ ...rest }
+				rowFormatter={ rowFormatter }
+				items={ items }
+				isLoading={ isLoading }
+				hasMoreActions={ hasMoreActions }
+				primaryKey={ primaryKey }
+			/>
 			<div className="plugins-list__mobile-view">
 				<>
-					{ rest.title && (
+					{ title && (
 						<Card className="plugins-list__content-header">
-							<div>{ rest.title }</div>
+							<div>{ title }</div>
 						</Card>
 					) }
 
-					{ rest.isLoading ? (
+					{ isLoading ? (
 						<Card>
 							<TextPlaceholder />
 						</Card>
 					) : (
-						rest.items.map( ( item ) => (
+						items.map( ( item ) => (
 							<PluginCard
 								{ ...rest }
+								selectedSite={ selectedSite }
+								rowFormatter={ rowFormatter }
 								key={ item[ primaryKey ] }
 								item={ item }
 								hasMoreActions={ hasMoreActions }
